+++ traveller-android-app/traveller/src/main/java/com/requestlabs/traveller/App.java
@@ -4,6 +4,7 @@ import android.app.Application;
 
 import com.parse.Parse;
 import com.parse.ParseAnalytics;
+import com.parse.ParseObject;
 
 /**
  * Created by gjones on 5/16/14.
@@ -16,6 +17,7 @@ public class App extends Application {
     public void onCreate()
     {
         super.onCreate();
+        ParseObject.registerSubclass(Item.class);
         requireParseConfiguration();
         Parse.initialize(this, Constants.api_key, Constants.client_id);
     }